PythonBackend/requirements.txt
openai>=1.0.0
llama-cpp-python[server]>=0.3.0
pyyaml>=6.0
uvicorn>=0.30.0
